refactor: consolidate telemetry into Nitro plugin
Move OTel setup (MeterProvider, PrometheusExporter, counters/histograms)
into plugins/telemetry.ts. Add custom 'chronicle:ssr-rendered' Nitro
hook so entry-server can record SSR metrics via useNitroApp().hooks.

Removes separate telemetry.ts module — all state lives in plugin scope.
