--- packages/chronicle/src/server/telemetry.ts
@@ -1,41 +1,0 @@
-import type { Counter, Histogram } from '@opentelemetry/api'
-import { MeterProvider } from '@opentelemetry/sdk-metrics'
-import { PrometheusExporter } from '@opentelemetry/exporter-prometheus'
-import { resourceFromAttributes } from '@opentelemetry/resources'
-import { ATTR_SERVICE_NAME } from '@opentelemetry/semantic-conventions'
-import type { ChronicleConfig } from '@/types/config'
-
-let exporter: PrometheusExporter
-let requestCounter: Counter
-let requestDuration: Histogram
-let ssrRenderDuration: Histogram
-
-export function initTelemetry(config: ChronicleConfig) {
-  const resource = resourceFromAttributes({
-    [ATTR_SERVICE_NAME]: config.telemetry?.serviceName ?? 'chronicle',
-  })
-
-  const port = config.telemetry?.port ?? 9090
-  exporter = new PrometheusExporter({ port })
-  const provider = new MeterProvider({ resource, readers: [exporter] })
-  const meter = provider.getMeter('chronicle')
-
-  requestCounter = meter.createCounter('http_server_request_total', {
-    description: 'Total HTTP requests',
-  })
-  requestDuration = meter.createHistogram('http_server_request_duration_ms', {
-    description: 'HTTP request duration in ms',
-  })
-  ssrRenderDuration = meter.createHistogram('http_server_ssr_render_duration_ms', {
-    description: 'SSR render duration in ms',
-  })
-}
-
-export function recordRequest(method: string, route: string, status: number, durationMs: number) {
-  requestCounter?.add(1, { method, route, status })
-  requestDuration?.record(durationMs, { method, route, status })
-}
-
-export function recordSSRRender(route: string, status: number, durationMs: number) {
-  ssrRenderDuration?.record(durationMs, { route, status })
-}
